# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ConversationIn": {
        "properties": {
          "document_scope": {
            "items": {
              "type": "string"
            },
            "title": "Document Scope",
            "type": "array"
          },
          "knowledge_base_scope": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Knowledge Base Scope",
            "type": "array"
          },
          "title": {
            "default": "New conversation",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "knowledge_base_scope"
        ],
        "title": "ConversationIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "KnowledgeBaseIn": {
        "properties": {
          "cloud_allowed": {
            "default": false,
            "title": "Cloud Allowed",
            "type": "boolean"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "graph_enabled": {
            "default": false,
            "title": "Graph Enabled",
            "type": "boolean"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "KnowledgeBaseIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Personal RAG",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/conversations": {
      "get": {
        "operationId": "list_conversations_api_v1_conversations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Conversations"
      },
      "post": {
        "operationId": "create_conversation_api_v1_conversations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Conversation"
      }
    },
...
    "/api/v1/knowledge-bases": {
      "get": {
        "operationId": "list_knowledge_bases_api_v1_knowledge_bases_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Knowledge Bases"
      },
      "post": {
        "operationId": "create_knowledge_base_api_v1_knowledge_bases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Knowledge Base"
      }
    },
...
  }
}